production/todo_project/requirements.txt
Django==3.1
django-filter==2.4
djangorestframework==3.12
djangorestframework-simplejwt
django-cors-headers==3.8
Markdown
graphene
graphene-django==2.15.0
drf-yasg==1.20
mixer
psycopg2-binary